--- src/LatestNews.module.css
@@ -1,116 +1,0 @@
-
-@import url('https://fonts.googleapis.com/css2?family=Signika+Negative&family=Spectral:wght@300&display=swap');
-@import url('https://cdn.jsdelivr.net/npm/bootstrap-icons@1.3.0/font/bootstrap-icons.css'); 
-
-body{
-    font-family: 'Signika Negative', sans-serif;
-}
-.Newshead{
-text-align: center;
-
-}
-.Newshead a {
-    font-size: 1.3rem;
-    text-decoration: none;
-    font-weight: 600;
-    color: rgb(146, 75, 218);
-}
-.Newshead h1{
-    font-size: 2.5rem;
-    color: rgb(5, 51, 51);
-}
-
-.newscontent {
-    width: 380px;
-    padding: 15px;
-    text-align: center;
-
-}
-.newscontent img{
-    height: 320px;
-    width: 320px;
-    border-radius: 2px ;
-    transition: .2s;
-}
-.newscontent img:hover{
-    transform: scale(1.1) ;
-
-}
-.newscontent i{
-    color: rgb(129, 55, 204) ;
-    font-size: 1.1rem;
-}
-.newscontent p{
-    text-decoration: none;
-    font-size: 1.2rem;
-    opacity: 75%;
-    color: rgb(14, 85, 85) ;
-
-}
-.newscontent a {
- color: rgb(5, 51, 51) ;
- text-decoration: none;
- font-size: 1.5;
-
-}
-.newscontent a:hover{
-    color: rgb(129, 146, 248);
-}
-.newscontent button{
-    color: rgb(113, 132, 238);
-    text-decoration: underline;
-    border: none;
-    background-color: white;
-    font-weight: 700;
-    font-size: 1.2rem;
-    cursor: pointer;
-    text-transform: capitalize;
-    transition: .3s ease-in-out;
-
-}
-.newscontent button:hover{
-    color: rgb(14, 85, 85);
-    text-decoration:none;
-}
-.Allnews{
-
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    flex-flow: wrap;
-
-}
-
-@media only screen and   (min-width:600px ) and  (max-width: 830px){
-    .newscontent{
-        width: 30%;
-       height: max-content;
-    }
-
-    .newscontent img{
-     width: 100%;
-     height: 100%;
-
-    }
-}
-
-
-@media only screen and  (max-width: 350px){
-.newscontent{
-   width: 100%;
-   height: max-content;
-
-
-}
-.newscontent img {
-   width: 180px;
-   height: 180px;
-}
-
-}
-
-.Newshead{
-
-    border-top: 5px solid rgb(227, 242, 243);
-
-}
