+++ src/index.js
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import ReactDOM from "react-dom";
 
-import LatestNews from "./LatestNews.js";
+import { LatestNews } from "./components";
 
 const rootElement = document.getElementById("root");
 ReactDOM.render(
